--- .spaceignore
@@ -1,11 +1,0 @@
-api/.venv/
-node_modules/
-build/
-.svelte-kit/
-.env.me
-
-*/.env
-*__pycache__
-*/a.ipynb
-src/
-tests/
