--- Spacefile
@@ -1,12 +1,0 @@
-v: 0
-icon: ./Icon.png
-app_name: 'drive'
-micros:
-  - name: drive-app
-    src: .
-    engine: svelte-kit
-    primary: true
-    public: true
-    presets:
-      env:
-        - name: JWT_SECRET
