+++ package.json
@@ -9,17 +9,17 @@
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "prettier --plugin-search-dir . --check .",
-    "format": "prettier --plugin-search-dir . --write .",
-    "local-test": "serve build"
+    "format": "prettier --plugin-search-dir . --write ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-node": "^1.3.1",
-    "@sveltejs/kit": "^1.30.4",
+    "@sveltejs/adapter-vercel": "^5.3.1",
+    "@sveltejs/kit": "^2.5.10",
+    "@sveltejs/vite-plugin-svelte": "^3.1.1",
     "@types/bcrypt": "^5.0.2",
     "@types/js-yaml": "^4.0.9",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/ms": "^0.7.34",
-    "@zerodevx/svelte-toast": "^0.8.2",
+    "@zerodevx/svelte-toast": "^0.9.5",
     "autoprefixer": "^10.4.19",
     "bcrypt": "^5.1.1",
     "eslint": "^8.57.0",
@@ -27,17 +27,18 @@
     "jsonwebtoken": "^9.0.2",
     "postcss": "^8.4.38",
     "postcss-load-config": "^4.0.2",
-    "prettier": "^2.8.8",
-    "prettier-plugin-svelte": "^2.10.1",
-    "prettier-plugin-tailwindcss": "^0.2.8",
+    "prettier": "^3.3.1",
+    "prettier-plugin-svelte": "^3.2.4",
+    "prettier-plugin-tailwindcss": "^0.6.2",
     "sass": "^1.77.4",
-    "svelte": "^3.59.2",
-    "svelte-check": "^2.10.3",
-    "svelte-preprocess": "^4.10.7",
-    "tailwindcss": "^3.4.3",
+    "svelte": "^4.2.18",
+    "svelte-check": "^3.8.0",
+    "svelte-preprocess": "^5.1.4",
+    "tailwindcss": "^3.4.4",
+    "trpc-sveltekit": "^3.6.1",
     "tslib": "^2.6.2",
-    "typescript": "^4.9.5",
-    "vite": "^4.5.3"
+    "typescript": "^5.4.5",
+    "vite": "^5.2.13"
   },
   "type": "module",
   "dependencies": {
@@ -45,7 +46,6 @@
     "@trpc/server": "^10.45.2",
     "ms": "^2.1.3",
     "svelte-icons-pack": "^2.1.1",
-    "trpc-sveltekit": "^3.6.1",
     "zod": "^3.23.8"
   }
 }
