--- src/langs/index.ts
@@ -1,2 +1,0 @@
-export * from './datt';
-export { default } from './datt';
